Improve iOS Demo app code organization
